* src/*: Backed off changes to namespaces and template classes with 	  Stream parameter, as they proved to be of no benefit to portability.
	* src/*: Finished off a bit more ISO conformance. (all std:: additions).

	* src/Message/Object.h: Copied changes from stable branch
	  including memory pool optimisation, asNum() and isNum() methods, and
	  change of ListType from std::list to std::vector.

	* src/*: Replaced int with long throughout code for compatability
	  with changes type in Message::Object

This updates the development version with much of the work done
on the stable branch since they split, but does not include the later
work done by sdt to get the factories system working on gcc 2.96 and later for
the Atlas-C++-0.4 release.

Al Riddoch (alriddoch@zepler.org)

+++ Atlas/Bridge.h
@@ -39,7 +39,7 @@ class Bridge
 
     virtual void mapItem(const std::string& name, const Map&) = 0;
     virtual void mapItem(const std::string& name, const List&) = 0;
-    virtual void mapItem(const std::string& name, int) = 0;
+    virtual void mapItem(const std::string& name, long) = 0;
     virtual void mapItem(const std::string& name, double) = 0;
     virtual void mapItem(const std::string& name, const std::string&) = 0;
     virtual void mapEnd() = 0;
@@ -48,7 +48,7 @@ class Bridge
     
     virtual void listItem(const Map&) = 0;
     virtual void listItem(const List&) = 0;
-    virtual void listItem(int) = 0;
+    virtual void listItem(long) = 0;
     virtual void listItem(double) = 0;
     virtual void listItem(const std::string&) = 0;
     virtual void listEnd() = 0;
